project/python-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
googlemaps==4.10.0
authlib==1.2.1
httpx==0.25.2
slowapi==0.1.9
limits==3.6.0
databases==0.8.0
python-decouple==3.8